requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.1
transformers==4.35.0
torch==2.3.0
sentence-transformers==2.2.2
faiss-cpu
orjson
pandas
pytest
python-multipart
requests
tqdm
gunicorn